Revert "Move RPC handlers to room (#599)"
This reverts commit db42022b2e99b7474b647e91823af1ae6a834f8f.
