+++ livekit-api/src/access_token.rs
@@ -282,10 +282,6 @@ impl TokenVerifier {
     pub fn verify(&self, token: &str) -> Result<Claims, AccessTokenError> {
         let mut validation = jsonwebtoken::Validation::new(jsonwebtoken::Algorithm::HS256);
         validation.validate_exp = true;
-        #[cfg(test)] // FIXME: TEST_TOKEN is expired, TODO: generate TEST_TOKEN at test runtime
-        {
-            validation.validate_exp = false;
-        }
         validation.validate_nbf = true;
         validation.set_issuer(&[&self.api_key]);
