+++ livekit-ffi/src/livekit.proto.rs
@@ -1,4 +1,5 @@
 // @generated
+// This file is @generated by prost-build.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct FrameCryptor {
@@ -4043,7 +4044,7 @@ pub struct PerformRpcRequest {
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct RegisterRpcMethodRequest {
     #[prost(uint64, required, tag="1")]
-    pub room_handle: u64,
+    pub local_participant_handle: u64,
     #[prost(string, required, tag="2")]
     pub method: ::prost::alloc::string::String,
 }
@@ -4051,7 +4052,7 @@ pub struct RegisterRpcMethodRequest {
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct UnregisterRpcMethodRequest {
     #[prost(uint64, required, tag="1")]
-    pub room_handle: u64,
+    pub local_participant_handle: u64,
     #[prost(string, required, tag="2")]
     pub method: ::prost::alloc::string::String,
 }
