+++ livekit-ffi/src/server/participant.rs
@@ -16,6 +16,7 @@ use std::sync::Arc;
 
 use livekit::prelude::*;
 use std::time::Duration;
+use tokio::sync::oneshot;
 
 use crate::{
     proto,
@@ -76,4 +77,88 @@ impl FfiParticipant {
         server.watch_panic(handle);
         Ok(proto::PerformRpcResponse { async_id })
     }
+
+    pub fn register_rpc_method(
+        &self,
+        server: &'static FfiServer,
+        request: proto::RegisterRpcMethodRequest,
+    ) -> FfiResult<proto::RegisterRpcMethodResponse> {
+        let method = request.method.clone();
+
+        let local = match &self.participant {
+            Participant::Local(local) => local.clone(),
+            Participant::Remote(_) => {
+                return Err(FfiError::InvalidRequest("Expected local participant".into()))
+            }
+        };
+
+        let local_participant_handle = self.handle.clone();
+        let room: Arc<RoomInner> = self.room.clone();
+        local.register_rpc_method(method.clone(), move |data| {
+            Box::pin({
+                let room = room.clone();
+                let method = method.clone();
+                async move {
+                    forward_rpc_method_invocation(
+                        server,
+                        room,
+                        local_participant_handle,
+                        method,
+                        data,
+                    )
+                    .await
+                }
+            })
+        });
+        Ok(proto::RegisterRpcMethodResponse {})
+    }
+
+    pub fn unregister_rpc_method(
+        &self,
+        request: proto::UnregisterRpcMethodRequest,
+    ) -> FfiResult<proto::UnregisterRpcMethodResponse> {
+        let local = match &self.participant {
+            Participant::Local(local) => local.clone(),
+            Participant::Remote(_) => {
+                return Err(FfiError::InvalidRequest("Expected local participant".into()))
+            }
+        };
+
+        local.unregister_rpc_method(request.method);
+
+        Ok(proto::UnregisterRpcMethodResponse {})
+    }
+}
+
+async fn forward_rpc_method_invocation(
+    server: &'static FfiServer,
+    room: Arc<RoomInner>,
+    local_participant_handle: FfiHandleId,
+    method: String,
+    data: RpcInvocationData,
+) -> Result<String, RpcError> {
+    let (tx, rx) = oneshot::channel();
+    let invocation_id = server.next_id();
+
+    room.store_rpc_method_invocation_waiter(invocation_id, tx);
+
+    let _ = server.send_event(proto::ffi_event::Message::RpcMethodInvocation(
+        proto::RpcMethodInvocationEvent {
+            local_participant_handle: local_participant_handle as u64,
+            invocation_id,
+            method,
+            request_id: data.request_id,
+            caller_identity: data.caller_identity.into(),
+            payload: data.payload,
+            response_timeout_ms: data.response_timeout.as_millis() as u32,
+        },
+    ));
+
+    rx.await.unwrap_or_else(|_| {
+        Err(RpcError {
+            code: RpcErrorCode::ApplicationError as u32,
+            message: "Error from method handler".to_string(),
+            data: None,
+        })
+    })
 }
